ostatni

+++ src/Controller/CategoryController.php
@@ -7,11 +7,12 @@ namespace App\Controller;
 
 use App\Entity\Category;
 use App\Form\CategoryType;
-use App\Repository\CategoryRepository;
-use Knp\Component\Pager\PaginatorInterface;
+use App\Service\CategoryService;
+use Doctrine\ORM\OptimisticLockException;
+use Doctrine\ORM\ORMException;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
-use Symfony\Component\Form\Extension\Core\Type\SearchType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -24,11 +25,26 @@ use Symfony\Component\Routing\Annotation\Route;
 class CategoryController extends AbstractController
 {
     /**
+     * Category service.
+     *
+     * @var \App\Service\CategoryService
+     */
+    private $categoryService;
+
+    /**
+     * CategoryController constructor.
+     *
+     * @param \App\Service\CategoryService $categoryService Category service
+     */
+    public function __construct(CategoryService $categoryService)
+    {
+        $this->categoryService = $categoryService;
+    }
+
+    /**
      * Index action.
      *
-     * @param \App\Repository\CategoryRepository            $repository Category repository
-     * @param \Knp\Component\Pager\PaginatorInterface   $paginator  Paginator
-     * @param \Symfony\Component\HttpFoundation\Request $request    HTTP request
+     * @param \Symfony\Component\HttpFoundation\Request $request HTTP request
      *
      * @return \Symfony\Component\HttpFoundation\Response HTTP response
      *
@@ -38,24 +54,21 @@ class CategoryController extends AbstractController
      *     name="category_index",
      * )
      */
-    public function index(CategoryRepository $repository, PaginatorInterface $paginator, Request $request): Response
+    public function index(Request $request): Response
     {
-        $pagination = $paginator->paginate(
-            $repository->queryAll(),
-            $request->query->getInt('page', 1),
-            Category::NUMBER_OF_ITEMS
-        );
-        $form = $this->createForm(SearchType::class);
-        $form->handleRequest($request);
+        $page = $request->query->getInt('page', 1);
+        $pagination = $this->categoryService->createPaginatedList($page);
 
-        return $this->render('category/index.html.twig', ['pagination' => $pagination, 'form' => $form->createView()]);
+        return $this->render(
+            'category/index.html.twig',
+            ['pagination' => $pagination]
+        );
     }
 
     /**
      * New action.
      *
-     * @param \Symfony\Component\HttpFoundation\Request $request    HTTP request
-     * @param \App\Repository\CategoryRepository            $repository category repository
+     * @param \Symfony\Component\HttpFoundation\Request $request HTTP request
      *
      * @return \Symfony\Component\HttpFoundation\Response HTTP response
      *
@@ -68,17 +81,14 @@ class CategoryController extends AbstractController
      *     name="category_new",
      * )
      */
-    public function new(Request $request, CategoryRepository $repository): Response
+    public function new(Request $request): Response
     {
         $category = new Category();
         $form = $this->createForm(CategoryType::class, $category);
 
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            $repository->save($category);
-            /*
-             * Potwierdzienie zapisania
-             */
+            $this->categoryService->save($category);
             $this->addFlash('success', 'message_created_successfully');
 
             return $this->redirectToRoute('category_index');
@@ -91,11 +101,10 @@ class CategoryController extends AbstractController
     }
 
     /**
-     * Delete action.
+     * Edit action.
      *
-     * @param \Symfony\Component\HttpFoundation\Request $request    HTTP request
-     * @param \App\Entity\Category                          $category   Category entity
-     * @param \App\Repository\CategoryRepository            $repository Category repository
+     * @param \Symfony\Component\HttpFoundation\Request $request  HTTP request
+     * @param \App\Entity\Category                      $category Category entity
      *
      * @return \Symfony\Component\HttpFoundation\Response HTTP response
      *
@@ -103,14 +112,60 @@ class CategoryController extends AbstractController
      * @throws \Doctrine\ORM\OptimisticLockException
      *
      * @Route(
+     *     "/{id}/edit",
+     *     methods={"GET", "PUT"},
+     *     requirements={"id": "[1-9]\d*"},
+     *     name="category_edit",
+     * )
+     */
+    public function edit(Request $request, Category $category): Response
+    {
+        $form = $this->createForm(CategoryType::class, $category, ['method' => 'PUT']);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->categoryService->save($category);
+            $this->addFlash('success', 'message_updated_successfully');
+
+            return $this->redirectToRoute('category_index');
+        }
+
+        return $this->render(
+            'category/edit.html.twig',
+            [
+                'form' => $form->createView(),
+                'category' => $category,
+            ]
+        );
+    }
+
+    /**
+     * Delete action.
+     *
+     * @param Request  $request  HTTP request
+     * @param Category $category Category entity
+     *
+     * @return Response HTTP response
+     *
+     * @throws ORMException
+     * @throws OptimisticLockException
+     *
+     * @Route(
      *     "/{id}/delete",
      *     methods={"GET", "DELETE"},
      *     requirements={"id": "[1-9]\d*"},
      *     name="category_delete",
      * )
+     * @IsGranted("ROLE_ADMIN")
      */
-    public function delete(Request $request, Category $category, CategoryRepository $repository): Response
+    public function delete(Request $request, Category $category): Response
     {
+        if ($category->getAnnouncements()->count()) {
+            $this->addFlash('warning', 'message_tag_contains_announcement');
+
+            return $this->redirectToRoute('category_index');
+        }
+
         $form = $this->createForm(FormType::class, $category, ['method' => 'DELETE']);
         $form->handleRequest($request);
 
@@ -119,8 +174,8 @@ class CategoryController extends AbstractController
         }
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $repository->delete($category);
-            $this->addFlash('success', 'message.deleted_successfully');
+            $this->categoryService->delete($category);
+            $this->addFlash('success', 'message_deleted_successfully');
 
             return $this->redirectToRoute('category_index');
         }
