+++ src/Controller/TagController.php
@@ -6,29 +6,44 @@
 namespace App\Controller;
 
 use App\Entity\Tag;
-use App\Form\TagType;
-use App\Repository\TagRepository;
-use Knp\Component\Pager\PaginatorInterface;
+use App\Service\TagService;
+use Doctrine\ORM\OptimisticLockException;
+use Doctrine\ORM\ORMException;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
-use Symfony\Component\Form\Extension\Core\Type\SearchType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * Class TagController.
+ * Class CategoryController.
  *
  * @Route("/tag")
  */
 class TagController extends AbstractController
 {
     /**
+     * Category service.
+     *
+     * @var \App\Service\TagService
+     */
+    private $tagService;
+
+    /**
+     * CategoryController constructor.
+     *
+     * @param \App\Service\TagService $tagService Category service
+     */
+    public function __construct(TagService $tagService)
+    {
+        $this->tagService = $tagService;
+    }
+
+    /**
      * Index action.
      *
-     * @param \App\Repository\TagRepository            $repository Tag repository
-     * @param \Knp\Component\Pager\PaginatorInterface   $paginator  Paginator
-     * @param \Symfony\Component\HttpFoundation\Request $request    HTTP request
+     * @param \Symfony\Component\HttpFoundation\Request $request HTTP request
      *
      * @return \Symfony\Component\HttpFoundation\Response HTTP response
      *
@@ -37,70 +52,29 @@ class TagController extends AbstractController
      *     methods={"GET"},
      *     name="tag_index",
      * )
+     * @IsGranted("ROLE_ADMIN")
      */
-    public function index(TagRepository $repository, PaginatorInterface $paginator, Request $request): Response
+    public function index(Request $request): Response
     {
-        $pagination = $paginator->paginate(
-            $repository->queryAll(),
-            $request->query->getInt('page', 1),
-            Tag::NUMBER_OF_ITEMS
-        );
-        $form = $this->createForm(SearchType::class);
-        $form->handleRequest($request);
-
-        return $this->render('tag/index.html.twig', ['pagination' => $pagination, 'form' => $form->createView()]);
-    }
-
-    /**
-     * New action.
-     *
-     * @param \Symfony\Component\HttpFoundation\Request $request    HTTP request
-     * @param \App\Repository\TagRepository            $repository tag repository
-     *
-     * @return \Symfony\Component\HttpFoundation\Response HTTP response
-     *
-     * @throws \Doctrine\ORM\ORMException
-     * @throws \Doctrine\ORM\OptimisticLockException
-     *
-     * @Route(
-     *     "/new",
-     *     methods={"GET", "POST"},
-     *     name="tag_new",
-     * )
-     */
-    public function new(Request $request, TagRepository $repository): Response
-    {
-        $tag = new Tag();
-        $form = $this->createForm(TagType::class, $tag);
-
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $repository->save($tag);
-            /*
-             * Potwierdzienie zapisania
-             */
-            $this->addFlash('success', 'message_created_successfully');
-
-            return $this->redirectToRoute('tag_index');
-        }
+        $page = $request->query->getInt('page', 1);
+        $pagination = $this->tagService->createPaginatedList($page);
 
         return $this->render(
-            'tag/new.html.twig',
-            ['form' => $form->createView()]
+            'tag/index.html.twig',
+            ['pagination' => $pagination]
         );
     }
 
     /**
      * Delete action.
      *
-     * @param \Symfony\Component\HttpFoundation\Request $request    HTTP request
-     * @param \App\Entity\Tag                          $tag   Tag entity
-     * @param \App\Repository\TagRepository            $repository Tag repository
+     * @param Request $request HTTP request
+     * @param Tag     $tag     Tag entity
      *
-     * @return \Symfony\Component\HttpFoundation\Response HTTP response
+     * @return Response HTTP response
      *
-     * @throws \Doctrine\ORM\ORMException
-     * @throws \Doctrine\ORM\OptimisticLockException
+     * @throws ORMException
+     * @throws OptimisticLockException
      *
      * @Route(
      *     "/{id}/delete",
@@ -108,9 +82,15 @@ class TagController extends AbstractController
      *     requirements={"id": "[1-9]\d*"},
      *     name="tag_delete",
      * )
+     * @IsGranted("ROLE_ADMIN")
      */
-    public function delete(Request $request, Tag $tag, TagRepository $repository): Response
+    public function delete(Request $request, Tag $tag): Response
     {
+        if ($tag->getAnnouncements()->count()) {
+            $this->addFlash('warning', 'message_tag_contains_announcement');
+
+            return $this->redirectToRoute('tag_index');
+        }
         $form = $this->createForm(FormType::class, $tag, ['method' => 'DELETE']);
         $form->handleRequest($request);
 
@@ -119,7 +99,7 @@ class TagController extends AbstractController
         }
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $repository->delete($tag);
+            $this->tagService->delete($tag);
             $this->addFlash('success', 'message.deleted_successfully');
 
             return $this->redirectToRoute('tag_index');
